Hoist predicate providers to root events dir
