+++ src/main/resources/META-INF/services/tech.thatgravyboat.skyblockapi.api.events.base.EventPredicateProvider
@@ -1,1 +1,1 @@
-com.github.mkram17.bazaarutils.events.screen.predicates.BazaarScreenEventPredicateProvider
+com.github.mkram17.bazaarutils.events.predicates.BazaarScreenEventPredicateProvider
